pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shade-product-catalog"
version = "0.1.0"
description = "Product catalog API (KMAT diagrams, BOM, published snapshots)"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "pypdf>=5.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.2",
    "httpx==0.28.1",
    "ruff==0.12.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py314"
preview = true

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
